send default values to client and add environment variables

+++ SlotMachineApiNetCore2/Config/Params.cs
@@ -4,7 +4,7 @@
     {
         public int MaxRows { get; set; }
         public int MaxCols { get; set; }
-        public double DefaultBetAmount { get; set; }
+        public int DefaultBetAmount { get; set; }
         public double PayoutRatio { get; set; }
         public double InitialBalance { get; set; }
         public int TimerInterval { get; set; }
